# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Callception API",
    "description": "AI-powered call management, appointment tracking, and customer service automation platform. Provides voice AI pipeline, CRM, ticketing, billing, and multi-tenant management capabilities.",
    "version": "1.0.0",
    "contact": {
      "name": "Callception",
      "url": "https://callception.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://callception.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3002",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "TenantHeader": []
    }
  ],
...
  "paths": {
...
    "/api/voice/pipeline": {
      "post": {
        "tags": ["Voice Pipeline"],
        "summary": "Start or interact with voice pipeline session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tenantId": { "type": "string" },
                  "sessionId": { "type": "string" },
                  "action": { "type": "string", "enum": ["start", "text"], "default": "start" },
                  "language": { "type": "string" },
                  "text": { "type": "string", "description": "Required for action=text" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Pipeline session info or text response" },
          "503": { "description": "Pipeline not configured" }
        }
      },
      "get": {
        "tags": ["Voice Pipeline"],
        "summary": "Pipeline health/configuration check",
        "responses": { "200": { "description": "Pipeline provider status" } }
      }
    },
...
    "/api/billing/emergency": {
      "get": {
        "tags": ["Billing"],
        "summary": "Get emergency mode status",
        "responses": { "200": { "description": "Emergency mode status and cost usage" } }
      },
      "post": {
        "tags": ["Billing"],
        "summary": "Toggle emergency mode",
        "description": "Emergency mode switches body TTS to cheaper OpenAI provider.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action"],
                "properties": {
                  "action": { "type": "string", "enum": ["activate", "deactivate", "auto", "update_config"] },
                  "config": { "type": "object", "description": "For action=update_config" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated emergency status" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase JWT",
        "description": "Firebase ID token for authenticated requests"
      },
      "TenantHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-tenant",
        "description": "Tenant ID header set by authentication middleware"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for automation/webhook endpoints"
      }
    },
...
  }
}